requirements.txt
PyQt6>=6.6.0
manga-ocr>=0.1.11
edge-tts>=6.1.0
mss>=9.0.0
requests>=2.31.0
keyboard>=0.13.5
Pillow>=10.0.0
numpy>=1.24.0
markdown>=3.5.0
